copy/requirements.txt
psycopg2-binary
python-dotenv==1.0.0
